4.11 集成持久层mybatis官方代码生成器 访问数据库成功

+++ member/src/main/java/com/anchor/train/member/controller/TestController.java
@@ -12,7 +12,7 @@ public class TestController {
     @Resource
     private MemberService memberService;
     @GetMapping("/count")
-    public Integer count() {
+    public long count() {
         return memberService.count();
     }
 }
